# contracts/generated/api.ts
/* generated by `bullet contracts generate`
 * source: contracts/openapi.yaml
 * regenerate: cargo run -p bullet -- contracts generate
 * do not hand-edit */

export type Health = {
  status: string;
};

export type ObservationKind = "value" | "empty" | "unknown" | "contradictory";

export type Mission = {
  id: string;
  organization_id: string;
  repository_id: string;
  title: string;
  objective: string;
  acceptance_contract_id: string;
  state: string;
};

export type WorkPackage = {
  id: string;
  mission_id: string;
  plan_revision_id: string;
  task_class: string;
  title: string;
  state: string;
};

export type MissionView = {
  mission: Mission;
  packages: WorkPackage[];
  fence: number | null;
};

export type DemoReceipt = {
  mission_id: string;
  plan_hash: string;
  fence: number;
  attempt_id: string;
  fence_second: number;
  attempt_second_id: string;
  stale_attempt_id: string;
  candidate_head: string;
  evidence_result: string;
  effect_outcome: string;
  effect_unknown_outcome: string;
  materialize_idempotent: boolean;
  stale_refused: boolean;
};

export type OutboxItem = {
  seq: number;
  kind: string;
  payload: string;
  phase: string;
  delivered_at: string | null;
  acked_at: string | null;
};

export type OutboxView = {
  items: OutboxItem[];
};

export type Event = {
  seq: number;
  kind: string;
  body: string;
  event_id: string | null;
  stream_id: string | null;
  sequence: number | null;
  causation_id: string | null;
  correlation_id: string | null;
  authority_token_hash: string | null;
};

export type Problem = {
  type: string;
  title: string;
  status: number;
  code: string;
  correlation_id: string;
  retryable: boolean;
};

export const API_PREFIX = "/v1";